Package processing and sending

+++ broker/submit.py
@@ -1,20 +1,34 @@
 from __future__ import annotations
 
-from threading import Thread
+import shutil
+import subprocess
+import sys
+from threading import Thread, Lock
 from pathlib import Path
 from enum import Enum
 from datetime import datetime
+from time import sleep
+
 from yaml import safe_load
 
 from baca2PackageManager import Package
 from db.connector import Connection
 from .builder import Builder
-from settings import BUILD_NAMESPACE
+from settings import BUILD_NAMESPACE, KOLEJKA_CONF
 
 from typing import TYPE_CHECKING
+from shlex import quote as shlex_quote
 
 if TYPE_CHECKING:
     from .master import BrokerMaster
+
+
+def _translate_paths(*args):
+    for arg in args:
+        if isinstance(arg, Path):
+            yield str(arg)
+        else:
+            yield arg
 
 
 class SubmitState(Enum):
@@ -44,6 +58,8 @@ class SubmitState(Enum):
 
 
 class TaskSubmit(Thread):
+    TIMEOUT_STEP = 3
+
     def __init__(self,
                  master: BrokerMaster,
                  submit_id: str,
@@ -51,6 +67,7 @@ class TaskSubmit(Thread):
                  commit_id: str,
                  submit_path: Path,
                  force_rebuild: bool = False,
+                 verbose: bool = False,
                  *args, **kwargs):
         super().__init__(*args, **kwargs)
         self.state = SubmitState.ADOPTING
@@ -60,32 +77,71 @@ class TaskSubmit(Thread):
         self.package_path = package_path
         self.commit_id = commit_id
         self.package = Package(self.package_path, self.commit_id)
+        self.build_path = None
         self.submit_path = submit_path
-        self.result_path = master.results_dir / submit_id
+        self.result_path = master.submits_dir / submit_id
         self.force_rebuild = force_rebuild
+        self.verbose = verbose
         self.sets = []
-
+        self._submit_update_lock = Lock()
+        self.sets_statuses = {
+            SubmitState.ADOPTING: 0,
+            SubmitState.SENDING: 0,
+            SubmitState.AWAITING_JUDGE: 0,
+            SubmitState.SAVING: 0,
+            SubmitState.DONE: 0,
+            SubmitState.ERROR: 0,
+        }
         self._conn = master.connection
         self._conn.exec("INSERT INTO submit_records VALUES (?, ?, ?, ?, ?, ?, NULL, ?)",
-                        (self.submit_id,  # id
-                         datetime.now(),  # launch_datetime
-                         self.submit_path,  # submit_path
-                         self.package_path,  # package_path
-                         self.commit_id,  # commit_id
-                         self.result_path,  # result_path
-                         self.state  # state
-                         ))
+                        self.submit_id,  # id
+                        datetime.now(),  # launch_datetime
+                        self.submit_path,  # submit_path
+                        self.package_path,  # package_path
+                        self.commit_id,  # commit_id
+                        self.result_path,  # result_path
+                        self.state  # state
+                        )
+        self.task_submit_dir = self.master.submits_dir / self.submit_id
+        if self.task_submit_dir.is_dir():
+            shutil.rmtree(self.task_submit_dir)
+        self.task_submit_dir.mkdir()
+
+    def vprint(self, msg: str):
+        if self.verbose:
+            print(f'<{self.submit_id}> {msg}')
 
     def _change_state(self, state: SubmitState, error_msg: str = None):
         self.state = state
-        if state == SubmitState.DONE:
-            if self.master.delete_records:
-                self._conn.exec("DELETE FROM submit_records WHERE id = ?", (self.submit_id,))
+        if state == SubmitState.DONE and self.master.delete_records:
+                self._conn.exec("DELETE FROM submit_records WHERE id = ?", self.submit_id, )
             # TODO: Check if following exits thread.
-            self.master.close_submit()
+            # self.master.close_submit()
         else:
             self._conn.exec("UPDATE submit_records SET state=?, error_msg=? WHERE id=?",
-                            (state, error_msg, self.submit_id))
+                            state, error_msg, self.submit_id)
+
+        if self.verbose:
+            self.vprint(f'Changed state to {state.value} ({state.name})')
+
+    @property
+    def kolejka_client(self) -> Path:
+        return self.package.build_path(BUILD_NAMESPACE) / 'common' / 'kolejka-client'
+
+    @property
+    def kolejka_judge(self) -> Path:
+        return self.package.build_path(BUILD_NAMESPACE) / 'common' / 'kolejka-judge'
+
+    @property
+    def judge_py(self) -> Path:
+        return self.package.build_path(BUILD_NAMESPACE) / 'common' / 'judge.py'
+
+    def set_submit_update(self, status: SubmitState):
+        self._submit_update_lock.acquire()
+        self.sets_statuses[status] += 1
+        self._submit_update_lock.release()
+        if self.sets_statuses[status] == len(self.sets):
+            self._change_state(status)
 
     def close_set_submit(self, set_name: str):
         pass
@@ -132,8 +188,17 @@ class TaskSubmit(Thread):
         for s in self.sets:
             s.start()
 
+    def run(self):
+        try:
+            self.process()
+        except Exception as e:
+            self._change_state(SubmitState.ERROR, str(e))
+
 
 class SetSubmit(Thread):
+    class KOLEJKACommunicationFailed(Exception):
+        pass
+
     def __init__(self,
                  master: BrokerMaster,
                  task_submit: TaskSubmit,
@@ -150,34 +215,106 @@ class SetSubmit(Thread):
         self.submit_id = submit_id
         self.package = package
         self.set_name = set_name
-        self.set_submit_url = f'{self.submit_id}_{self.set_name}'
+        self.callback_url = None
 
         self.submit_path = submit_path
-        self.result_path = master.results_dir / submit_id / set_name
         self._conn = master.connection
         self.submit_http_server = master.submit_http_server
 
         self._conn.exec("INSERT INTO set_submit_records VALUES (NULL, ?, ?, NULL, ?)",
-                        (self.submit_id, self.set_name, self.state))
+                        self.submit_id, self.set_name, self.state)
+        self.task_dir = self.task_submit.task_submit_dir / f'{self.set_name}.task'
+        self.result_dir = self.task_submit.task_submit_dir / f'{self.set_name}.result'
+
+        if sys.platform.startswith('win'):
+            self.python_call = 'py'
+        else:
+            self.python_call = 'python3'
+
+        # self._call_for_update()
+
+    def vprint(self, msg: str):
+        self.task_submit.vprint(f'[{self.set_name}] {msg}')
+
+    def _call_for_update(self):
+        self.task_submit.set_submit_update(self.state)
 
     def _change_state(self, state: SubmitState, error_msg: str = None):
         # TODO: Add state monitoring for parent (task) submit
         self.state = state
-        if state == SubmitState.DONE:
-            if self.master.delete_records:
-                self._conn.exec("DELETE FROM set_submit_records WHERE submit_id=? AND set_name=?",
-                                (self.submit_id, self.set_name))
+        if state == SubmitState.DONE and self.master.delete_records:
+            self._conn.exec("DELETE FROM set_submit_records WHERE submit_id=? AND set_name=?",
+                            self.submit_id, self.set_name)
             # TODO: Check if following exits thread.
+        elif state == SubmitState.DONE or state == SubmitState.ERROR:
             self.task_submit.close_set_submit(self.set_name)
         else:
             self._conn.exec("UPDATE set_submit_records SET state=?, error_msg=? WHERE submit_id=? AND set_name=?",
-                            (state, error_msg, self.submit_id, self.set_name))
+                            state, error_msg, self.submit_id, self.set_name)
+        if self.task_submit.verbose:
+            self.vprint(f'Changed state to {state.value} ({state.name})')
+        self._call_for_update()
 
     def _send_submit(self):
-        self.submit_http_server.add_submit(self.set_submit_url)
+        self.callback_url = self.submit_http_server.add_submit(f'{self.submit_id}_{self.set_name}')
+        # TODO: adding submit should return url
 
-    def _await_results(self, timeout: float = -1) -> bool:
-        return self.submit_http_server.await_submit(self.set_submit_url, timeout)
+        cmd_judge = [self.python_call, self.task_submit.kolejka_judge,
+                     'task',
+                     '--callback', self.callback_url,
+                     '--library-path', self.task_submit.kolejka_judge,
+                     self.task_submit.judge_py,
+                     self.package.build_path(BUILD_NAMESPACE) / self.set_name / "tests.yaml",
+                     self.submit_path,
+                     self.task_dir]
+        cmd_client = [self.python_call, self.task_submit.kolejka_client,
+                      '--config-file', KOLEJKA_CONF,
+                      'task', 'put',
+                      self.task_dir]
+
+        # kolejka-client result get <result_dir>
+
+        cmd_judge = list(_translate_paths(*cmd_judge))
+        cmd_client = list(_translate_paths(*cmd_client))
+
+        judge_status = subprocess.run(cmd_judge)
+        if judge_status.returncode != 0:
+            raise self.KOLEJKACommunicationFailed('KOLEJKA judge failed to create task.')
+
+        client_status = subprocess.run(cmd_client, capture_output=True)
+        self.result_code = client_status.stdout.decode('utf-8').strip()
+
+        if client_status.returncode != 0:
+            raise self.KOLEJKACommunicationFailed('KOLEJKA client failed to communicate with KOLEJKA server.')
+
+    def _await_results(self, timeout: float = -1, active_wait:bool = False) -> bool:
+        if not active_wait:
+            self.submit_http_server.await_submit(self.set_submit_url, timeout)
+
+        result_get = [self.python_call,
+                      self.task_submit.kolejka_client,
+                        '--config-file', KOLEJKA_CONF,
+                      'result', 'get',
+                      self.result_code,
+                      self.result_dir]
+
+        result_get = list(_translate_paths(*result_get))
+
+        result_status = subprocess.run(result_get, stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL)
+
+        if not active_wait:
+            if result_status.returncode != 0:
+                raise self.KOLEJKACommunicationFailed('KOLEJKA client failed to communicate with KOLEJKA server.')
+        else:
+            time_track = 0
+            while result_status.returncode != 0:
+                sleep(self.task_submit.TIMEOUT_STEP)
+                time_track += self.task_submit.TIMEOUT_STEP
+                result_status = subprocess.run(result_get, stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL)
+                if 0 < timeout < time_track:
+                    raise self.KOLEJKACommunicationFailed('Reached timeout while waiting for KOLEJKA response.')
+
+        return True
 
     def process(self):
         self._change_state(SubmitState.SENDING)
@@ -185,10 +322,9 @@ class SetSubmit(Thread):
 
         self._change_state(SubmitState.AWAITING_JUDGE)
         # TODO: consider adding timeout for safety (otherwise if something goes wrong the method may never return)
-        self._await_results(timeout=-1)
+        self._await_results(timeout=60, active_wait=True)
 
         self._change_state(SubmitState.SAVING)
-        self._call_for_update(success=True)
 
         self._change_state(SubmitState.DONE)
